+++ stroom-util/src/main/java/stroom/util/io/GZipConstants.java
@@ -1,0 +1,8 @@
+package stroom.util.io;
+
+public final class GZipConstants {
+    /**
+     * Default size (in bytes) for GZIP I/O stream buffers
+     */
+    public static final int DEFAULT_BUFFER_SIZE = 64 * 1024;
+}