+++ stroom-util/src/main/java/stroom/util/io/WrappedOutputStream.java
@@ -20,7 +20,7 @@ import java.io.IOException;
 import java.io.OutputStream;
 
 public class WrappedOutputStream extends OutputStream {
-    private final OutputStream outputStream;
+    protected final OutputStream outputStream;
 
     public WrappedOutputStream(final OutputStream outputStream) {
         this.outputStream = outputStream;