+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 * Issue **#1866** : Change zip distribution shell scripts to execute from anywhere.
 
-## [v6.2.0 - Unreleased]
+* Issue **#759** : Add GZIP support for `FileAppender` and `RollingFileAppender`.
 
 * Add Elasticsearch integration
 
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Add common secret property encryption utils (`CryptoUtils`)
 
 * Add common property `stroom.secret.encryptionKey` for encrypting secrets at rest
+
 
 ## [v6.1.32] - 2021-01-28